Move network init into separate unit
